Job_Posting/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0  
anthropic>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
httpx>=0.24.0